requirements.txt
# Main requirements
python-telegram-bot[job-queue]==20.7
SQLAlchemy==2.0.25
requests==2.31.0
toml==0.10.2
python-dotenv==1.0.0

# Sub-dependencies
APScheduler==3.10.4
certifi==2023.11.17
chardet==5.2.0
coloredlogs==15.0.1
greenlet>=3.0.3
humanfriendly==10.0
idna==3.6
pytz==2023.3.post1
six==1.16.0
tornado==6.4
tzlocal==5.2
urllib3==2.1.0
